src/python/requirements.txt
aniso8601==9.0.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
dnspython==2.1.0
Flask==2.0.0
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycodestyle==2.7.0
pymongo==3.11.4
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
toml==0.10.2
urllib3==1.26.4
Werkzeug==2.0.1